+++ docs/deprecated/sockets/index.js
@@ -1,5 +1,6 @@
 import { initCancellationToken } from "./_cancellationToken.js";
 import { fishTalesSocket } from "./_fishTalesSocket.js";
+import { shopReelSocket } from "./_shopReelSocket.js";
 import { siteScoutSocket } from "./_siteScoutSocket.js";
 
-export { initCancellationToken, fishTalesSocket, siteScoutSocket };
+export { initCancellationToken, fishTalesSocket, shopReelSocket, siteScoutSocket };